requirements.txt
Markdown==3.0.1
djangorestframework==3.9.0
Flask==1.0.2
dj-database-url==0.5.0
Django>=2.2
pylint==2.3.1
autopep8==1.4.4